requirements.txt
Flask==3.1.2
Flask-MySQLdb==2.0.0
pandas==3.0.0
nltk==3.9.2
sentence-transformers==5.2.3
scikit-learn==1.8.0
language-tool-python==3.2.2
mysql-connector-python==9.6.0